+++ src/Backends/Rendering/Window/OpenGL3/SDL2.cpp
@@ -15,7 +15,7 @@ SDL_Window *window;
 
 static SDL_GLContext context;
 
-bool WindowBackend_OpenGL_CreateWindow(const char *window_title, int *screen_width, int *screen_height, bool fullscreen)
+bool WindowBackend_OpenGL_CreateWindow(const char *window_title, size_t *screen_width, size_t *screen_height, bool fullscreen)
 {
 #ifdef USE_OPENGLES2
 	if (SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_ES) < 0)
