+++ src/Backends/Rendering/Window/Software/Null.cpp
@@ -6,7 +6,7 @@
 static unsigned char *framebuffer;
 static size_t framebuffer_pitch;
 
-bool WindowBackend_Software_CreateWindow(const char *window_title, int screen_width, int screen_height, bool fullscreen)
+bool WindowBackend_Software_CreateWindow(const char *window_title, size_t screen_width, size_t screen_height, bool fullscreen)
 {
 	(void)window_title;
 	(void)fullscreen;
@@ -45,7 +45,7 @@ void WindowBackend_Software_Display(void)
 	
 }
 
-void WindowBackend_Software_HandleWindowResize(unsigned int width, unsigned int height)
+void WindowBackend_Software_HandleWindowResize(size_t width, size_t height)
 {
 	(void)width;
 	(void)height;
